+++ 2lab/POS/src/main/java/com/team1206/pos/inventory/product/ProductResponseDTO.java
@@ -5,12 +5,13 @@ import lombok.Data;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
+import java.math.BigDecimal;
 
 @Data
 public class ProductResponseDTO {
     private UUID id;                // Unique identifier for the product
     private String name;            // Product name
-    private Integer price;          // Product price in smallest currency unit (e.g., cents)
+    private BigDecimal price;       // Product price in smallest currency unit (e.g., cents)
     private UUID categoryId;        // ID of the associated ProductCategory
     private List<UUID> variationIds; // List of associated ProductVariation IDs
     private List<UUID> chargeIds;   // List of associated Charge IDs
